requirements.txt
easydict
h5py
nltk
numpy
Pillow
torch
torchvision
tqdm
termcolor
pyyaml
streamlit
flask
flask-restful